Move fixtures to a folder that makes more sense to contain them
