+++ src/App.tsx
@@ -1,4 +1,3 @@
-import { PokemonItem } from "./components/PokemonItem/";
 import { PokemonList } from "./components/PokemonList/";
 
 function App() {
