+++ src/fixtures/axios/get/pokemon/fixture.ts
